Revert "Add test for prebundled apps"
This change appears to cause a crash according to James Donald:
I suspected that my Airspace crashes might be tied to the installtype,
but it crashes even on a non-Pongo installation. Although the change is
labled 'add test for prebundled apps' it makes changes elsewhere and in
fact the crash is in somewhere for regular app scanning.

Based on the error, this one-line change can remove the crash :

diff --git a/app/models/local-leap-app.js b/app/models/local-leap-app.js
index 6472532..abe11ed 100644
--- a/app/models/local-leap-app.js
+++ b/app/models/local-leap-app.js
@@ -219,7 +219,8 @@ function localAppScan(manifest, cb) {
         oldApp.save();
       });
     }
-    cb && cb(null);
+    // FIXME: Removing this to avoid a 'callback was already called'
error
+    // cb && cb(null);
   });

This reverts commit 2ae3c05eb8c63b4f8c45493f6fc300283604a06d.